+++ ui/components/app/security-trust/index.ts
@@ -2,7 +2,4 @@ export {
   getSecurityTrustBadgeConfig,
   SecurityTrustInlineBadge,
 } from './security-trust-inline-badge';
-export type {
-  SecurityTrustInlineBadgeConfig,
-  SecurityTrustTranslate,
-} from './security-trust-inline-badge';
+export type { SecurityTrustTranslate } from './security-trust-inline-badge';
